Height Change Channel - Thickness adjustment
This channel contains the latest 2D profile with other enhancements intended for the rest of the channels. This commit offers a thickness slider (in advanced settings). The thickness slider adjusts the thickness of the top and chamfer sections of the top channel while preserving the thickness of the clamp structure and total width of the channel, and available space inside.